requirements.txt
# ===================================================================
# DEPENDENCIAS DEL PROYECTO LSP (Lengua de Señas Peruana)
# ===================================================================
# Compatibilidad: Python 3.10, 3.11, 3.12
# Última actualización: Octubre 2025
#
# OPCIONES DE INSTALACIÓN:
#   1. Versión actualizada (recomendada): pip install -r requirements.txt
#   2. Versión legacy compatible: Ver sección "Legacy" al final
# ===================================================================

# ===================================================================
# CORE - Deep Learning Framework
# ===================================================================
tensorflow==2.15.0
# Nota: Keras viene incluido en TensorFlow 2.15+ (no instalar por separado)
# Si tienes problemas, considera: tensorflow==2.13.0

# ===================================================================
# COMPUTER VISION
# ===================================================================
mediapipe==0.10.11
# MediaPipe para detección de pose, manos y rostro
# Versión compatible: 0.8.11 (si hay problemas con 0.10.11)

opencv-contrib-python==4.10.0.84
# OpenCV con módulos contrib para procesamiento de video
# Alternativa más ligera: opencv-python==4.10.0.84

# ===================================================================
# PROCESAMIENTO DE DATOS
# ===================================================================
numpy==1.26.4
# NumPy para operaciones numéricas
# Compatible con TensorFlow 2.15

pandas==2.2.2
# Pandas para manejo de DataFrames

tables==3.9.2
# PyTables para almacenamiento HDF5 (archivos .h5 de keypoints)

# ===================================================================
# INTERFAZ DE USUARIO
# ===================================================================
PyQt5==5.15.11
# PyQt5 para la interfaz gráfica (GUI)
# Requiere Qt5 en el sistema

# ===================================================================
# TEXT-TO-SPEECH Y AUDIO
# ===================================================================
gTTS==2.5.3
# Google Text-to-Speech para síntesis de voz

pygame==2.6.1
# Pygame para reproducción de audio

# ===================================================================
# WEB FRAMEWORK (Servidor API REST)
# ===================================================================
Flask==3.0.3
# Flask para servidor web

Werkzeug==3.0.3
# Dependencia de Flask (manejo de requests)

# ===================================================================
# UTILIDADES
# ===================================================================
PyYAML==6.0.2
# Para cargar archivo de configuración config.yaml

protobuf==3.20.3
# Protocol Buffers (requerido por MediaPipe y TensorFlow)
# IMPORTANTE: MediaPipe 0.10.11 requiere protobuf<4.0

# ===================================================================
# MACHINE LEARNING TOOLS (Opcional pero recomendado)
# ===================================================================
scikit-learn==1.5.1
# Para métricas de evaluación y validación cruzada

matplotlib==3.9.2
# Para visualizaciones y gráficos

seaborn==0.13.2
# Para gráficos estadísticos mejorados (matriz de confusión)

# ===================================================================
# TESTING (Opcional - Solo para desarrollo)
# ===================================================================
pytest==8.3.2
# Framework de testing

pytest-cov==5.0.0
# Cobertura de tests


# ===================================================================
# INSTALACIÓN ALTERNATIVA - VERSIONES LEGACY
# ===================================================================
# Si encuentras problemas con las versiones actualizadas,
# crea un archivo requirements_legacy.txt con estas versiones:
#
# tensorflow==2.8.0
# keras==2.8.0
# mediapipe==0.8.11
# numpy==1.21.6
# opencv-python==4.5.5.64
# pandas==1.4.4
# pygame==2.5.2
# Flask==3.0.2
# protobuf==3.19.1
# tables==3.7.0
# PyQt5==5.15.9
# gTTS==2.5.1
# PyYAML==6.0.2
#
# Instalar con: pip install -r requirements_legacy.txt
# ===================================================================


# ===================================================================
# NOTAS DE INSTALACIÓN
# ===================================================================
#
# 1. CREAR ENTORNO VIRTUAL (Recomendado):
#    python -m venv venv
#    .\venv\Scripts\Activate.ps1  # Windows PowerShell
#    source venv/bin/activate      # Linux/Mac
#
# 2. INSTALAR DEPENDENCIAS:
#    pip install --upgrade pip
#    pip install -r requirements.txt
#
# 3. VERIFICAR INSTALACIÓN:
#    python -c "import tensorflow as tf; print(f'TensorFlow: {tf.__version__}')"
#    python -c "import mediapipe; print('MediaPipe: OK')"
#    python -c "from config_manager import config; print('Config: OK')"
#
# 4. PROBLEMAS COMUNES:
#
#    a) Error con protobuf:
#       pip install protobuf==3.20.3 --force-reinstall
#
#    b) Error con MediaPipe en Python 3.12:
#       Usar Python 3.10 o 3.11 (MediaPipe tiene mejor soporte)
#
#    c) Error con TensorFlow en GPU:
#       pip install tensorflow[and-cuda]==2.15.0
#
#    d) Error con PyQt5:
#       Instalar Qt5: https://www.qt.io/download
#
#    e) Conflictos de versiones:
#       pip install --upgrade --force-reinstall -r requirements.txt
#
# 5. PARA CONTRIBUIDORES:
#    pip install -e .
#    pip install -r requirements.txt
#    pip install black flake8 mypy  # Herramientas de linting
#
# ===================================================================